Use int16_t for entity numbers.

+++ src/net/net_message.c
@@ -757,7 +757,7 @@ void Net_ReadDeltaPlayerState(mem_buf_t *msg, const player_state_t *from, player
 
   *to = *from;
 
-  const int32_t bits = Net_ReadShort(msg);
+  const uint16_t bits = Net_ReadShort(msg);
 
   if (bits & PS_PM_CLIENT) {
     to->client = Net_ReadByte(msg);
@@ -828,7 +828,7 @@ void Net_ReadDeltaPlayerState(mem_buf_t *msg, const player_state_t *from, player
  * @brief
  */
 void Net_ReadDeltaEntity(mem_buf_t *msg, const entity_state_t *from, entity_state_t *to,
-                         uint16_t number, uint16_t bits) {
+                         int16_t number, uint16_t bits) {
 
   *to = *from;
